first small changes to epidural/subdural insertion
